Migrate models.Recipe.search -> search.RecipeSearch.query
